Added tests for Jupiter-Saturn system, printout of solar system RHS, and A^2 + L^2 = const test.
The solar system values don't agree well with the values from the
Murray and Dermott appendix, but I'm not sure how much those values
are influenced by non-secular perturbations.

Also, even for the secular Jupiter-Saturn system example in M&D, the
agreement with raw-average is only in the first significant digit.